+++ Scripts/Cards/RelentlessPursuit.cs
@@ -1,0 +1,55 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using BaseLib.Abstracts;
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.HoverTips;
+using MegaCrit.Sts2.Core.Localization;
+using MegaCrit.Sts2.Core.Localization.DynamicVars;
+using MegaCrit.Sts2.Core.Models.CardPools;
+using USCE.Scripts.Powers;
+
+namespace USCE.Scripts.Cards;
+
+[Pool(typeof(SilentCardPool))]
+public class RelentlessPursuit : SilentCardModel, ILocalizationProvider
+{
+    private const int energyCost = 1;
+    private const CardType type = CardType.Power;
+    private const CardRarity rarity = CardRarity.Uncommon;
+    private const TargetType targetType = TargetType.Self;
+
+    protected override IEnumerable<DynamicVar> CanonicalVars =>
+    [
+        new PowerVar<RelentlessPursuitPower>(4m)
+    ];
+
+    public override List<(string, string)>? Localization => LocManager.Instance.Language switch
+    {
+        "zhs" => new CardLoc("穷追不舍", "打出攻击牌时，如果你的[gold]手牌[/gold]中有与其名字相同的牌，额外造成{RelentlessPursuitPower:diff()}点伤害。"),
+        _ => new CardLoc("Relentless Pursuit", "When you play an Attack, if you have a card with the same name in your [gold]hand[/gold], deal {RelentlessPursuitPower:diff()} extra damage.")
+    };
+
+    public RelentlessPursuit() : base(energyCost, type, rarity, targetType)
+    {
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay cardPlay)
+    {
+        await CreatureCmd.TriggerAnim(Owner.Creature, "Cast", Owner.Character.CastAnimDelay);
+        if (IsUpgraded)
+        {
+            await PowerCmd.Apply<RelentlessPursuitPowerPlus>(Owner.Creature, 6m, Owner.Creature, this);
+        }
+        else
+        {
+            await PowerCmd.Apply<RelentlessPursuitPower>(Owner.Creature, 4m, Owner.Creature, this);
+        }
+    }
+
+    protected override void OnUpgrade()
+    {
+    }
+}